+++ requires_io/consts.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+REQUIREMENTS = 'requirements'
+SETUP = 'setup.py'
+TOX = 'tox'
+BUILDOUT = 'buildout'
+TYPES = (REQUIREMENTS, SETUP, TOX, BUILDOUT)